refactor: 예외 수정 및 페이지 컨트롤러 분리
