+++ src/main/java/roomescape/controller/ReservationPageController.java
@@ -1,0 +1,18 @@
+package roomescape.controller;
+
+import org.springframework.stereotype.Controller;
+import org.springframework.web.bind.annotation.GetMapping;
+
+@Controller
+public class ReservationPageController {
+
+    @GetMapping("/admin")
+    public String home() {
+        return "admin/index";
+    }
+
+    @GetMapping("/admin/reservation")
+    public String reservationPage() {
+        return "admin/reservation-legacy";
+    }
+}
